% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_cand.R
\name{choose_cand}
\alias{choose_cand}
\title{Selects candidates from a list}
\usage{
choose_cand(df, firstCandidate = NULL, secondCandidate = NULL)
}
\arguments{
\item{df}{the list of candidates}

\item{firstCandidate}{The number of the first candidate in the list, if known (otherwise will prompt)}

\item{secondCandidate}{The number of the second candidate in the list, if known}
}
\description{
Selects candidates from a list
}
