% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_get_fec.R
\name{get_fec}
\alias{get_fec}
\title{A FEC Scrape Function}
\usage{
get_fec(state, candidates)
}
\arguments{
\item{state:}{specifies state to scrape data}

\item{candidates:}{specifies which candidates to scrape}
}
\description{
THis function allows you to scrape data from the FEC.
}
\examples{
get_fec()
}
\keyword{fec,}
\keyword{politics}
