% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_avg_donation.R
\name{plot_avg_donation}
\alias{plot_avg_donation}
\title{Plot the average donation for candidates over the time window}
\usage{
plot_avg_donation(df)
}
\arguments{
\item{df}{A dataframe from query_contributions_all which specifies individual donations to candidates}
}
\description{
Plot the average donation for candidates over the time window
}
