% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cum_donation.R
\name{plot_cum_donation}
\alias{plot_cum_donation}
\title{Plot the cummulative donation for candidates over the time window}
\usage{
plot_cum_donation(df)
}
\arguments{
\item{df}{A dataframe from query_contributions_all which specifies individual donations to candidates}
}
\description{
Plot the cummulative donation for candidates over the time window
}
