% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_occupations.R
\name{plot_occupations}
\alias{plot_occupations}
\title{Plot occputation level data for donations}
\usage{
plot_occupations(n, df)
}
\arguments{
\item{n}{number of occpuations (max 8)}

\item{df}{A dataframe from query_contributions_all which specifies individual donations to candidates}
}
\description{
Plot occputation level data for donations
}
