% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top_cities.R
\name{plot_top_cities}
\alias{plot_top_cities}
\title{Plot zipcode level data for cities with the most donations}
\usage{
plot_top_cities(n, df)
}
\arguments{
\item{n}{number of top-contribution cities desired}

\item{df}{A dataframe from query_contributions_all which specifies individual donations to candidates}
}
\description{
Plot zipcode level data for cities with the most donations
}
