% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_candidate_list.R
\name{query_candidate_list}
\alias{query_candidate_list}
\title{Search for political candidates in FEC}
\usage{
query_candidate_list(api_key = NULL, state = NULL,
  election_year = NULL, office = NULL, candidate_status = "C")
}
\arguments{
\item{api_key}{An API key required to use OpenFEC}

\item{state}{A two-letter acronym which specifies the state to search (e.g., "TX")}

\item{election_year}{The year the election occured}

\item{office}{Specifies which office the candidate ran for: "S" = Senate, "H" = House, "G" = Governor, "P" = President}

\item{candidate_status}{Specifies whether the candidate is active during that election. Defaults to "C" for current.}
}
\description{
This function constructs searches for candidates listed in the FEC based on input parameters.
}
