% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_contributions_all.R
\name{query_contributions_all}
\alias{query_contributions_all}
\title{Find all individual donations to candidate}
\usage{
query_contributions_all(input_candlist, api_key)
}
\arguments{
\item{input_candlist}{Requires a df with candidate name and committee ID. Use query_candidate.R to get this information.}

\item{api_key}{API key for https://api.data.gov. Get one at https://api.data.gov/signup.}
}
\description{
This function constructs searches for donations listed in the FEC based on a candidate input
}
