% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_openfec.R
\name{query_openfec}
\alias{query_openfec}
\title{Make a request to the OpenFEC API}
\usage{
query_openfec(path, query_parameters = list())
}
\arguments{
\item{path}{A character string of the OpenFEC endpoint.}

\item{query_parameters}{A list with values of the parameters to use in the request.}
}
\description{
This function constructs the correct URL to send a GET request to the OpenFEC API. Returns the results a parsed JSON list. OpenFEC query parameters are documented here https://api.open.fec.gov/developers/.
}
